KVM: x86/mmu: handle cmpxchg failure in kvm_tdp_mmu_get_root
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1945742

To reduce dependence on the MMU write lock, don't rely on the assumption
that the atomic operation in kvm_tdp_mmu_get_root will always succeed.
By not relying on that assumption, threads do not need to hold the MMU
lock in write mode in order to take a reference on a TDP MMU root.

In the root iterator, this change means that some roots might have to be
skipped if they are found to have a zero refcount. This will still never
happen as of this patch, but a future patch will need that flexibility to
make the root iterator safe under the MMU read lock.

Signed-off-by: Ben Gardon <bgardon@google.com>
Message-Id: <20210401233736.638171-8-bgardon@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit fb10129335ca6cc7a229226b03f54394757d773d)
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>