crypto: ccp - Add support for SEV-ES to the PSP driver
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1945742

To provide support for SEV-ES, the hypervisor must provide an area of
memory to the PSP. Once this Trusted Memory Region (TMR) is provided to
the PSP, the contents of this area of memory are no longer available to
the x86.

Update the PSP driver to allocate a 1MB region for the TMR that is 1MB
aligned and then provide it to the PSP through the SEV INIT command.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>
Reviewed-by: Brijesh Singh <brijesh.singh@amd.com>
Reviewed-by: Joerg Roedel <jroedel@suse.de>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 97f9ac3db6612f14ac0c509e1a63ce14fd4cc0eb)
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>