vmxnet3: Set the default of vxlan overlay offload to disabled
Bugzilla: https://bugzilla.redhat.com/1941714
Upstream Status: RHEL only

This patch was supplied by Vmware engineering as a RHEL-only patch.

As part of RHEL 8.3 commit 45d7aa964 'vmxnet3: add geneve and vxlan tunnel offload support'
the VXLAN overlay offload feature was set to enabled by default.

When VXLAN offload is enabled, the vmxnet3 driver assumes it can offload
the checksumming of all packets sent through VXLAN tunnels, but the underlying
virtual NIC actually only performs offloaded VXLAN checksumming for VXLAN
packets that use a specific destination port. As a result, when doing VXLAN on
a port other than the one that VMware expects, the kernel will send packets
with invalid checksums to the NIC (because it thinks that the NIC will fill in
the checksums itself), the NIC will pass those packets along unchanged (because
it doesn't think it's supposed to do VXLAN offloading for packets on that port),
and eventually the packets containing the bad checksums will be dropped. This
resulted in upgrade failures for several OpenShift customers and the OpenShift
cluster would go offline. A request was made by OpenShift to change the default
of the overlay offload feature from enabled to disabled. Customers who wish to enable
the overlay offload feature themselves can do so.

Suggested-by:  Ronak Doshi <doshir@vmware.com>
Signed-off-by: Cathy Avery <cavery@redhat.com>