Amends commit ea6244cc248b to include a version string in module info.
Commit ea6244cc248b causes modinfo to return a null string for the
driver version. This commit fixes that so that the driver version
field is filled in with UTS_RELEASE.

V2: Instead of reverting completely, just use UTS_RELEASE for the driver
version to be consistent with other drivers.

Upstream Status: RHEL-only
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1955726
Tested: locally tested to verify that modinfo does not return a
	null driver version string

Signed-off-by: Ken Cox <jkc@redhat.com>