redhat/configs: Update torture test configs
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1941211
Upstream: RHEL only

With the latest RCU and KCSAN update, the following config changes
are introduced:
 - a new torture test for smp_call_function() is added
   (CONFIG_SCF_TORTURE_TEST)
 - a new CONFIG_KCSAN
 - CONFIG_RCU_PERF_TEST is renamed to CONFIG_RCU_SCALE_TEST

This patch enables the build of the new scftorture modules and adds it
to kernel-modules-internal. The rcutorture is still built only for the
debug kernel as enabling it for the generic kernel will break kABI.
CONFIG_KCSAN and CONFIG_RCU_SCALE_TEST are disabled.

Signed-off-by: Waiman Long <longman@redhat.com>