block: add capacity field to zone descriptors
Bugzilla: http://bugzilla.redhat.com/1921591

commit 82394db7383d33641f3f565bd79792fb41b1741f
Author: Matias Bjørling <matias.bjorling@wdc.com>
Date:   Mon Jun 29 12:06:37 2020 -0700

block: add capacity field to zone descriptors

In the zoned storage model, the sectors within a zone are typically all
writeable. With the introduction of the Zoned Namespace (ZNS) Command
Set in the NVM Express organization, the model was extended to have a
specific writeable capacity.

Extend the zone descriptor data structure with a zone capacity field to
indicate to the user how many sectors in a zone are writeable.

Introduce backward compatibility in the zone report ioctl by extending
the zone report header data structure with a flags field to indicate if
the capacity field is available.

Reviewed-by: Jens Axboe <axboe@kernel.dk>
Reviewed-by: Javier González <javier.gonz@samsung.com>
Reviewed-by: Chaitanya Kulkarni <chaitanya.kulkarni@wdc.com>
Reviewed-by: Himanshu Madhani <himanshu.madhani@oracle.com>
Reviewed-by: Martin K. Petersen <martin.petersen@oracle.com>
Reviewed-by: Hannes Reinecke <hare@suse.de>
Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Reviewed-by: Daniel Wagner <dwagner@suse.de>
Signed-off-by: Matias Bjørling <matias.bjorling@wdc.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>

Conflicts: Added KABI changes. 

./include/linux/rh_kabi.h:385:3: error: static assertion failed: "include/uapi/linux/blkzoned.h:114: struct { __u8 resv[4]; __u64 capacity; __u8 reserved[24]; } is larger than __u8 reserved[36].  Disable CONFIG_RH_KABI_SIZE_ALIGN_CHECKS if debugging."
_Static_assert(sizeof(struct{_new;}) <= sizeof(struct{_orig;}), \

It is triggered because padding will be added between 'u8 resv[4]' and
'__u64 capacity' by gcc, that is just our rh kabi helpers can't have the top
structure(struct blk_zone) info. The issue can be workaround simply by
__GENKSYMS__ wrapper.

Signed-off-by: Gopal Tiwari <gtiwari@redhat.com>