KVM: arm64: Ensure I-cache isolation between vcpus of a same VM
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1939323

It recently became apparent that the ARMv8 architecture has interesting
rules regarding attributes being used when fetching instructions
if the MMU is off at Stage-1.

In this situation, the CPU is allowed to fetch from the PoC and
allocate into the I-cache (unless the memory is mapped with
the XN attribute at Stage-2).

If we transpose this to vcpus sharing a single physical CPU,
it is possible for a vcpu running with its MMU off to influence
another vcpu running with its MMU on, as the latter is expected to
fetch from the PoU (and self-patching code doesn't flush below that
level).

In order to solve this, reuse the vcpu-private TLB invalidation
code to apply the same policy to the I-cache, nuking it every time
the vcpu runs on a physical CPU that ran another vcpu of the same
VM in the past.

This involve renaming __kvm_tlb_flush_local_vmid() to
__kvm_flush_cpu_context(), and inserting a local i-cache invalidation
there.

Cc: stable@vger.kernel.org
Signed-off-by: Marc Zyngier <maz@kernel.org>
Acked-by: Will Deacon <will@kernel.org>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Link: https://lore.kernel.org/r/20210303164505.68492-1-maz@kernel.org
(cherry picked from commit 01dc9262ff5797b675c32c0c6bc682777d23de05)
Signed-off-by: Gavin Shan <gshan@redhat.com>
Conflicts:
	arch/arm64/include/asm/kvm_asm.h
	arch/arm64/kvm/arm.c
	arch/arm64/kvm/hyp/nvhe/hyp-main.c
	arch/arm64/kvm/hyp/nvhe/tlb.c
	arch/arm64/kvm/hyp/vhe/tlb.c
Conflicts due to lack of upstream commits. The parameter passed to
__kvm_flush_cpu_context() is kept as vCPU instead of kvm_s2_mmu.
Changes related to hyp-main.c are dropped.
a0e50aa3f4a8 ("KVM: arm64: Factor out stage 2 page table data from struct kvm")
054698316d87 ("KVM: arm64: nVHE: Migrate hyp interface to SMCCC")