perf stat: Add --quiet option
Bugzilla: https://bugzilla.redhat.com/1944684

upstream
========
commit 55a4de94c64bacffbcd802c954764e0de2ab217f
Author: Andi Kleen <andi@firstfloor.org>
Date: Mon Oct 26 17:27:36 2020 -0700

description
===========
Add a new --quiet option to 'perf stat'. This is useful with 'perf stat
record' to write the data only to the perf.data file, which can lower
measurement overhead because the data doesn't need to be formatted.

On my 4C desktop:

  % time ./perf stat record  -e $(python -c 'print ",".join(["cycles"]*1000)')  -a -I 1000 sleep 5
  ...
  real    0m5.377s
  user    0m0.238s
  sys     0m0.452s
  % time ./perf stat record --quiet -e $(python -c 'print ",".join(["cycles"]*1000)')  -a -I 1000 sleep 5

  real    0m5.452s
  user    0m0.183s
  sys     0m0.423s

In this example it cuts the user time by 20%. On systems with more cores
the savings are higher.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>