perf mem: Search event name with more flexible path
Bugzilla: https://bugzilla.redhat.com/1944684

upstream
========
commit f9f16dfbe76e63ba9aec68055c08242b09be297e
Author: Leo Yan <leo.yan@linaro.org>
Date: Fri Nov 6 17:48:45 2020 +0800

description
===========
The perf tool searches a memory event name under the folder
'/sys/devices/cpu/events/', this leads to the limitation for the
selection of a memory profiling event which must be under this folder.

Thus it's impossible to use any other event as memory event which is not
under this specific folder, e.g. Arm SPE hardware event is not located
in '/sys/devices/cpu/events/' so it cannot be enabled for memory
profiling.

This patch changes to search folder from '/sys/devices/cpu/events/' to
'/sys/devices', so it give flexibility to find events which can be used
for memory profiling.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>