perf arm-spe: Refactor packet header parsing
Bugzilla: https://bugzilla.redhat.com/1944684

upstream
========
commit 11695142e25e957dc3e56c29dc5f9daaf9530b10
Author: Leo Yan <leo.yan@linaro.org>
Date: Thu Nov 19 23:24:27 2020 +0800

description
===========
The packet header parsing uses the hard coded values and it uses nested
if-else statements.

To improve the readability, this patch refactors the macros for packet
header format so it removes the hard coded values.  Furthermore, based
on the new mask macros it reduces the nested if-else statements and
changes to use the flat conditions checking, this is directive and can
easily map to the descriptions in ARMv8-a architecture reference manual
(ARM DDI 0487E.a), chapter 'D10.1.5 Statistical Profiling Extension
protocol packet headers'.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>