perf arm_spe: Fixup top byte for data virtual address
Bugzilla: https://bugzilla.redhat.com/1944684

upstream
========
commit 5513ddaf103c62dd1eabe9403c0a8d9f810492dc
Author: Leo Yan <leo.yan@linaro.org>
Date: Thu Nov 19 23:24:30 2020 +0800

description
===========
To establish a valid address from the address packet payload and finally
the address value can be used for parsing data symbol in DSO, current
code uses 0xff to replace the tag in the top byte of data virtual
address.

So far the code only fixups top byte for the memory layouts with 4KB
pages, it misses to support memory layouts with 64KB pages.

This patch adds the conditions for checking bits [55:52] are 0xf, if
detects the pattern it will fill 0xff into the top byte of the address,
also adds comment to explain the fixing up.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>