perf symbols: Try to load vmlinux from buildid database
Bugzilla: https://bugzilla.redhat.com/1944684

upstream
========
commit ca8ea73ae109900cec4c3a1f0d3486a01a0e4434
Author: Jiri Olsa <jolsa@kernel.org>
Date: Thu Nov 26 18:00:13 2020 +0100

description
===========
Currently we don't check on kernel's vmlinux the same way as we do for
normal binaries, but we either look for kallsyms file in build id
database or check on known vmlinux locations (plus some other optional
paths).

This patch adds the check for standard build id binary location, so we
are ready once we start to store it there from debuginfod in following
changes.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>