perf cpumap: Add new map type for aggregation
Bugzilla: https://bugzilla.redhat.com/1944684

upstream
========
commit cea6575fdccfc0624ca42f656e16e6b4d9bb48a5
Author: James Clark <james.clark@arm.com>
Date: Thu Nov 26 16:13:21 2020 +0200

description
===========
Currently this is a duplicate of perf_cpu_map so that it can be used as
a drop in replacement.

In a later commit it will be changed from a map of ints to use the new
cpu_aggr_id struct.

No functional changes.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>