perf stat: Introduce struct runtime_stat_data
Bugzilla: https://bugzilla.redhat.com/1944684

upstream
========
commit 3ff1e7180abc7f6db413933c110df69157216715
Author: Namhyung Kim <namhyung@kernel.org>
Date: Fri Jan 15 16:11:38 2021 +0900

description
===========
To pass more info to the saved_value in the runtime_stat, add a new
struct runtime_stat_data.  Currently it only has 'ctx' field but later
patch will add more.

Note that we intentionally pass 0 as ctx to clock-related events for
compatibility.  It was already there in a few places.  So move the code
into the saved_value_lookup() explicitly and add a comment.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>