perf unwind: Set userdata for all __report_module() paths
Bugzilla: https://bugzilla.redhat.com/1944684

upstream
========
commit 4e1481445407b86a483616c4542ffdc810efb680
Author: Dave Rigby <d.rigby@me.com>
Date:   Thu Feb 18 16:56:54 2021 +0000

description
===========
When locating the DWARF module for a given address, __find_debuginfo()
requires a 'struct dso' passed via the userdata argument.

However, this field is only set in __report_module() if the module is
found in via dwfl_addrmodule(), not if it is found later via
dwfl_report_elf().

Set userdata irrespective of how the DWARF module was found, as long as
we found a module.

Fixes: bf53fc6b5f41 ("perf unwind: Fix separate debug info files when using elfutils' libdw's unwinder")
Link: https://lore.kernel.org/linux-perf-users/20210218165654.36604-1-d.rigby@me.com/

Signed-off-by: Michael Petlan <mpetlan@redhat.com>