ALSA: hda: ignore invalid NHLT table
On some Lenovo systems if the microphone is disabled in the BIOS
only the NHLT table header is created, with no data. This means
the endpoints field is not correctly set to zero - leading to an
unintialised variable and hence invalid descriptors are parsed
leading to page faults.

The Lenovo firmware team is addressing this, but adding a check
preventing invalid tables being parsed is worthwhile.

Tested on a Lenovo T14.

    Tested-by: Philipp Leskovitz <philipp.leskovitz@secunet.com>
    Reported-by: Philipp Leskovitz <philipp.leskovitz@secunet.com>
    Signed-off-by: Mark Pearson <markpearson@lenovo.com>
    Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
    Link: https://lore.kernel.org/r/20210302141003.7342-1-markpearson@lenovo.com
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Mark Pearson <markpearson@lenovo.com>
Date: Tue Mar 2 09:10:03 2021 -0500

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit a14a6219996ee6f6e858d83b11affc7907633687)
Bugzilla: https://bugzilla.redhat.com/1917388