ALSA: hda/realtek: Add quirk for Intel NUC 10
This adds a new SND_PCI_QUIRK(...) and applies it to the Intel NUC 10
devices. This fixes the issue of the devices not having audio input and
output on the headset jack because the kernel does not recognize when
something is plugged in.

The new quirk was inspired by the quirk for the Intel NUC 8 devices, but
it turned out that the NUC 10 uses another pin. This information was
acquired by black box testing likely pins.

    Co-developed-by: Eckhart Mohr <e.mohr@tuxedocomputers.com>
    Signed-off-by: Eckhart Mohr <e.mohr@tuxedocomputers.com>
    Signed-off-by: Werner Sembach <wse@tuxedocomputers.com>
    Cc: <stable@vger.kernel.org>
    Link: https://lore.kernel.org/r/20210302180414.23194-1-wse@tuxedocomputers.com
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Werner Sembach <wse@tuxedocomputers.com>
Date: Tue Mar 2 19:04:14 2021 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 73e7161eab5dee98114987239ec9c87fe8034ddb)
Bugzilla: https://bugzilla.redhat.com/1917388